# Fix Performance/StringIdentifierArgument offense