--- a/.rubocop_todo.yml
+++ b/.rubocop_todo.yml
@@ -136,11 +136,6 @@ Performance/RegexpMatch:
     - 'lib/axlsx/workbook/worksheet/cell.rb'
 
 # This cop supports safe autocorrection (--autocorrect).
-Performance/StringIdentifierArgument:
-  Exclude:
-    - 'lib/axlsx/drawing/pic.rb'
-
-# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: separated, grouped
 Style/AccessorGrouping: